# Do not copy values to temporary memory buffer when inserting values into dictionaries