--- a/column_buffer_amd64.go
+++ b/column_buffer_amd64.go
@@ -3,10 +3,9 @@
 package parquet
 
 import (
-	"unsafe"
-
 	"github.com/segmentio/parquet-go/internal/bytealg"
 	"github.com/segmentio/parquet-go/internal/unsafecast"
+	"github.com/segmentio/parquet-go/sparse"
 	"golang.org/x/sys/cpu"
 )
 
@@ -18,7 +17,7 @@ func broadcastValueInt32(dst []int32, src int8) {
 func broadcastRangeInt32AVX2(dst []int32, base int32)
 
 func broadcastRangeInt32(dst []int32, base int32) {
-	if len(dst) >= minLenAVX2 && cpu.X86.HasAVX2 {
+	if len(dst) >= 8 && cpu.X86.HasAVX2 {
 		broadcastRangeInt32AVX2(dst, base)
 	} else {
 		for i := range dst {
@@ -28,52 +27,4 @@ func broadcastRangeInt32(dst []int32, base int32) {
 }
 
 //go:noescape
-func writeValuesBitpackAVX2(values unsafe.Pointer, rows array, size, offset uintptr)
-
-//go:noescape
-func writeValues32bitsAVX2(values unsafe.Pointer, rows array, size, offset uintptr)
-
-//go:noescape
-func writeValues64bitsAVX2(values unsafe.Pointer, rows array, size, offset uintptr)
-
-//go:noescape
-func writeValues128bits(values unsafe.Pointer, rows array, size, offset uintptr)
-
-func writeValuesBool(values []byte, rows array, size, offset uintptr) {
-	writeValuesBitpackAVX2(*(*unsafe.Pointer)(unsafe.Pointer(&values)), rows, size, offset)
-}
-
-func writeValuesInt32(values []int32, rows array, size, offset uintptr) {
-	writeValues32bitsAVX2(*(*unsafe.Pointer)(unsafe.Pointer(&values)), rows, size, offset)
-}
-
-func writeValuesInt64(values []int64, rows array, size, offset uintptr) {
-	writeValues64bitsAVX2(*(*unsafe.Pointer)(unsafe.Pointer(&values)), rows, size, offset)
-}
-
-func writeValuesUint32(values []uint32, rows array, size, offset uintptr) {
-	writeValues32bitsAVX2(*(*unsafe.Pointer)(unsafe.Pointer(&values)), rows, size, offset)
-}
-
-func writeValuesUint64(values []uint64, rows array, size, offset uintptr) {
-	writeValues64bitsAVX2(*(*unsafe.Pointer)(unsafe.Pointer(&values)), rows, size, offset)
-}
-
-func writeValuesFloat32(values []float32, rows array, size, offset uintptr) {
-	writeValues32bitsAVX2(*(*unsafe.Pointer)(unsafe.Pointer(&values)), rows, size, offset)
-}
-
-func writeValuesFloat64(values []float64, rows array, size, offset uintptr) {
-	writeValues64bitsAVX2(*(*unsafe.Pointer)(unsafe.Pointer(&values)), rows, size, offset)
-}
-
-func writeValuesBE128(values [][16]byte, rows array, size, offset uintptr) {
-	writeValues128bits(*(*unsafe.Pointer)(unsafe.Pointer(&values)), rows, size, offset)
-}
-
-//go:noescape
-func writePointers128bits(values *[16]byte, rows array, size, offset uintptr)
-
-func writePointersBE128(values [][16]byte, rows array, size, offset uintptr) {
-	writePointers128bits(*(**[16]byte)(unsafe.Pointer(&values)), rows, size, offset)
-}
+func writePointersBE128(values [][16]byte, rows sparse.Array)