--- a/hashprobe/aeshash/aeshash_amd64.go
+++ b/hashprobe/aeshash/aeshash_amd64.go
@@ -3,7 +3,7 @@
 package aeshash
 
 import (
-	"github.com/segmentio/parquet-go/hashprobe/sparse"
+	"github.com/segmentio/parquet-go/sparse"
 	"golang.org/x/sys/cpu"
 )